master: don't use $cron_time to determine whether or not cron should be used
The $cron_time parameter can legitimately be false. This tells the
puppet::cron class that it should automatically generate a random (based
on fqdn) time to run puppet within the hour.

Because of this, we shouldn't be using this variable to determine
whether or not the user wants to use cron or not. Currently, if
cron_time is false, the "puppet" class is used and this leads to an
error about the class being defined in two places.

Set usage of cron to be the default, since this is how it's recommended.